git-svn-id: https://svn.kwarc.info/repos/MMT@1529 09069c07-60fa-4895-8316-727242e252e4

+++ deploy/run-file.sh
@@ -1,2 +1,7 @@
 dir=`dirname $0`
-/usr/lib/jvm/java-7-openjdk-amd64/bin/java -Xmx2048m -cp $dir/lib/*:$dir/lfcatalog/*:$dir/mmt/* info.kwarc.mmt.api.frontend.Run file $1
+if [ -e "/usr/lib/jvm/java-7-openjdk-amd64/bin/java" ]; then
+  /usr/lib/jvm/java-7-openjdk-amd64/bin/java -Xmx2048m -cp $dir/lib/*:$dir/lfcatalog/*:$dir/mmt/* info.kwarc.mmt.api.frontend.Run file $1
+else
+    echo 'calling default java'
+    java -Xmx2048m -cp $dir/lib/*:$dir/lfcatalog/*:$dir/mmt/* info.kwarc.mmt.api.frontend.Run file $1
+fi
